requirements.txt
streamlit
pandas
numpy
torch
transformers
plotly
requests
python-dateutil
python-dotenv